@@ -1,23 +1,22 @@
 module github.com/alrusov/auth-url
 
-go 1.17
+go 1.19
 
 require (
-	github.com/alrusov/auth v0.1.0
-	github.com/alrusov/config v0.1.45
-	github.com/alrusov/log v0.1.31
-	github.com/alrusov/misc v0.1.49
-	github.com/alrusov/stdhttp v0.1.99
+	github.com/alrusov/auth v0.1.2
+	github.com/alrusov/config v0.1.47
+	github.com/alrusov/log v0.1.33
+	github.com/alrusov/misc v1.0.65
+	github.com/alrusov/stdhttp v0.1.106
 )
 
 require (
-	github.com/alrusov/bufpool v0.1.7 // indirect
-	github.com/alrusov/jsonw v0.1.1 // indirect
-	github.com/alrusov/loadavg v0.1.4 // indirect
-	github.com/alrusov/panic v0.1.14 // indirect
-	github.com/json-iterator/go v1.1.10 // indirect
+	github.com/alrusov/jsonw v0.1.3 // indirect
+	github.com/alrusov/loadavg v0.1.5 // indirect
+	github.com/alrusov/panic v0.1.15 // indirect
+	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
-	github.com/modern-go/reflect2 v0.0.0-20180701023420-4b7aa43c6742 // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/naoina/go-stringutil v0.1.0 // indirect
 	github.com/naoina/toml v0.1.1 // indirect
 )
